--- a/src/randgraphs.jl
+++ b/src/randgraphs.jl
@@ -50,6 +50,31 @@ function erdos_renyi(n::Integer, p::Real; is_directed=false)
     return g
 end
 
+
+using Distributions
+#faster for sparse erdos renyi graphs if p≈1, then use the erdos_renyi function
+function sparse_erdos_renyi(n::Integer, p::Real; is_directed=false)
+    if is_directed
+        possibleedges = n*(n - 1)
+        g = DiGraph(n)
+    else
+        possibleedges = (n^2-n)/2
+        g = Graph(n)
+    end
+    numedges = rand(Binomial(Int(possibleedges), p))
+    vtxdist  = DiscreteUniform(1, n)
+    k = 0
+    while k < numedges
+        i, j = rand(vtxdist), rand(vtxdist)
+        if !has_edge(g, i, j)
+            add_edge!(g, i, j)
+            k += 1
+        end
+    end
+    @assert ne(g)==numedges "the number of edges is wrong"
+    return g
+end
+
 function watts_strogatz(n::Integer, k::Integer, β::Real; is_directed=false)
     @assert k < n/2
     if is_directed