Here is a first draft it is already faster than the current n^2 implementation.
~~~~
julia> for n = [10^i for i =3:5]
       p = 200/(n*(n-1)); @time g = sparse_erdos_renyi(n,p); @show ne(g)
       end
elapsed time: 0.000176027 seconds (166136 bytes allocated)
ne(g) = 106
elapsed time: 0.000846324 seconds (1177832 bytes allocated)
ne(g) = 102
elapsed time: 0.008833541 seconds (11257336 bytes allocated)
ne(g) = 99

julia> for n = [10^i for i =3:5]
       p = 200/(n*(n-1)); @time g = erdos_renyi(n,p); @show ne(g)
       end
elapsed time: 0.002039418 seconds (153112 bytes allocated)
ne(g) = 86
elapsed time: 0.159099273 seconds (1172680 bytes allocated)
ne(g) = 108
elapsed time: 15.888211903 seconds (11252360 bytes allocated)
ne(g) = 106
~~~~
The following should be nearly constant time but isn't. I would guess because of garbage collection on growing the incidence lists.
~~~~julia
for n = [10^i for i =3:7]
      p = 2000/(n*(n-1));
      @time g = sparse_erdos_renyi(n,p)
      @show ne(g)
end
~~~~
Is that a problem with add_edge!