@sbromberger 2015-03-26
Three things, I guess:

1) why is this failing on -release?
2) could you make a quick change to the docs (I left a note at the right line)?
3) could you squash the commits?

Thanks.